+++ client/shared/src/search/integration/index.ts
@@ -1,0 +1,1 @@
+export * from './streaming-search-mocks'
